feat: Add GAScore benchmarking
feat: Update net_out to use TDEST
feat: Unify HLS and MB libraries into Active Messages
feat: Add Wireshark dissector for AMs
fix: Update sonar_strToInt to use new bitfields
refactor: Use custom typedefs in test software

The testbenches still need to be updated to use bytes
instead of words for payload size (same needs to be
reflected in the IPs as well). The SW library is distinct
in thegasnet* files but the MB and HLS are unified. The
core is also almost complete with the byte change needed
to be propogated through