Add support for Galapagos CPP library for SW
Change from word based sizes to bytes in packets
Resize strided message bitfield sizes for alignment
Update GAScore to optionally use absolute payloads
Update GAScore to use bitfield macros more fully
Change GAScore testbenches to use abs. payloads
Add in progress SW example with Gal. library
Add in prog. SW library for AM using Gal. library
Move some headers over to deprecated prior to deletion
fix: Some header misalingment in strToInt()

In the previous commit, I aligned the MB and HLS libraries. The SW one,
based on Reudi's original wouldn't mesh nicely with them. Naif's new
libray for Galapagos fits well but requires a rewrite of the Gasnet
connection. Here, I take a few parts of it to create a handler thread,
and discard the rest. It's incomplete at the moment but it's a work in
progress. The change from words to bytes was needed to transmit unaligned
data. However, this has not been tested at all since all the test cases
are still aligned. I also changed the payload size semantics to refer
to the absolute size of the packet instead of just the payload because
I thought I'd need to know the size of the packet in order to read it.
However, with Naif's library, this becomes moot so I should switch back
to the old semantics which are simpler to use.